Refactor for big refactoring
