+++ tests/creation/test_degraders.py
@@ -4,11 +4,12 @@ from typing import Type
 import numpy as np
 import pandas as pd
 import pytest
-from photerr import LsstErrorModel
 
 from rail.core.data import DATA_STORE, TableHandle
 from rail.core.utilStages import ColumnMapper
-from rail.creation.degradation import *
+from rail.creation.degradation.lsst_error_model import LSSTErrorModel
+from rail.creation.degradation.quantityCut import QuantityCut
+from rail.creation.degradation.spectroscopic_selections import *
 
 
 @pytest.fixture
@@ -47,86 +48,6 @@ def data_forspec():
     # return data in handle wrapping a pandas DataFrame
     df = pd.DataFrame(x, columns=["redshift", "u", "g", "r", "i", "z", "y"])
     return DS.add_data("data_forspec", df, TableHandle, path="dummy_forspec.pd")
-
-
-@pytest.mark.parametrize(
-    "true_wavelen,wrong_wavelen,frac_wrong,errortype",
-    [
-        ("fake", 200, 0.5, TypeError),
-        (100, "fake", 0.5, TypeError),
-        (100, 200, "fake", TypeError),
-        (-1, 200, 0.5, ValueError),
-        (100, -1, 0.5, ValueError),
-        (100, 200, -1, ValueError),
-        (100, 200, 2, ValueError),
-    ],
-)
-def test_LineConfusion_bad_params(true_wavelen, wrong_wavelen, frac_wrong, errortype):
-    """Test bad parameters that should raise ValueError"""
-    with pytest.raises(errortype):
-        LineConfusion.make_stage(
-            true_wavelen=true_wavelen,
-            wrong_wavelen=wrong_wavelen,
-            frac_wrong=frac_wrong,
-        )
-
-
-def test_LineConfusion_returns_correct_shape(data):
-    """Make sure LineConfusion doesn't change shape of data"""
-
-    degrader = LineConfusion.make_stage(name="xx", true_wavelen=100, wrong_wavelen=200, frac_wrong=0.5)
-    degraded_data = degrader(data).data
-
-    assert degraded_data.shape == data.data.shape
-    os.remove(degrader.get_output(degrader.get_aliased_tag("output"), final_name=True))
-
-
-def test_LineConfusion_no_negative_redshifts(data):
-    """Make sure that LineConfusion never returns a negative redshift"""
-
-    degrader = LineConfusion.make_stage(true_wavelen=100, wrong_wavelen=200, frac_wrong=0.5)
-    degraded_data = degrader(data).data
-
-    assert all(degraded_data["redshift"].to_numpy() >= 0)
-    os.remove(degrader.get_output(degrader.get_aliased_tag("output"), final_name=True))
-
-
-@pytest.mark.parametrize("pivot_redshift,errortype", [("fake", TypeError), (-1, ValueError)])
-def test_InvRedshiftIncompleteness_bad_params(pivot_redshift, errortype):
-    """Test bad parameters that should raise ValueError"""
-    with pytest.raises(errortype):
-        InvRedshiftIncompleteness.make_stage(pivot_redshift=pivot_redshift)
-
-
-def test_InvRedshiftIncompleteness_returns_correct_shape(data):
-    """Make sure returns same number of columns, fewer rows"""
-    degrader = InvRedshiftIncompleteness.make_stage(pivot_redshift=1.0)
-    degraded_data = degrader(data).data
-    assert degraded_data.shape[0] < data.data.shape[0]
-    assert degraded_data.shape[1] == data.data.shape[1]
-    os.remove(degrader.get_output(degrader.get_aliased_tag("output"), final_name=True))
-
-
-@pytest.mark.parametrize(
-    "percentile_cut,redshift_cut,errortype",
-    [(-1, 1, ValueError), (101, 1, ValueError), (99, -1, ValueError)],
-)
-def test_GridSelection_bad_params(percentile_cut, redshift_cut, errortype):
-    """Test bad parameters that should raise ValueError"""
-    with pytest.raises(errortype):
-        GridSelection.make_stage(percentile_cut=percentile_cut, redshift_cut=redshift_cut)
-
-
-def test_GridSelection_returns_correct_shape(data):
-    import pdb
-
-    """Make sure returns 2 more columns, fewer rows"""
-    degrader = GridSelection.make_stage(pessimistic_redshift_cut=1.0)
-    degraded_data = degrader(data).data
-    # pdb.set_trace()
-    assert degraded_data.shape[0] < data.data.shape[0]
-    assert degraded_data.shape[1] == data.data.shape[1] - 1
-    os.remove(degrader.get_output(degrader.get_aliased_tag("output"), final_name=True))
 
 
 @pytest.mark.parametrize(
@@ -290,8 +211,6 @@ def test_LSSTErrorModel_bad_type(errType, errortype):
 @pytest.mark.parametrize(
     "degrader",
     [
-        LineConfusion.make_stage(true_wavelen=100, wrong_wavelen=200, frac_wrong=0.01),
-        InvRedshiftIncompleteness.make_stage(pivot_redshift=1.0),
         LSSTErrorModel.make_stage(),
         LSSTErrorModel.make_stage(highSNR=True),
     ],
@@ -403,148 +322,3 @@ def test_SpecSelection_bad_path(success_rate_dir, errortype):
         SpecSelection.make_stage(success_rate_dir=success_rate_dir)
 
 
-def test_ObsCondition_returns_correct_shape(data):
-    """Test that the ObsCondition returns the correct shape"""
-
-    degrader = ObsCondition.make_stage()
-
-    degraded_data = degrader(data).data
-
-    assert degraded_data.shape == (data.data.shape[0], 2 * data.data.shape[1])
-    os.remove(degrader.get_output(degrader.get_aliased_tag("output"), final_name=True))
-
-
-def test_ObsCondition_random_seed(data):
-    """Test control with random seeds."""
-    degrader1 = ObsCondition.make_stage(random_seed=0)
-    degrader2 = ObsCondition.make_stage(random_seed=0)
-
-    # make sure setting the same seeds yields the same output
-    degraded_data1 = degrader1(data).data
-    degraded_data2 = degrader2(data).data
-    assert degraded_data1.equals(degraded_data2)
-
-    # make sure setting different seeds yields different output
-    degrader3 = ObsCondition.make_stage(random_seed=1)
-    degraded_data3 = degrader3(data).data.to_numpy()
-    assert not degraded_data1.equals(degraded_data3)
-
-    os.remove(degrader3.get_output(degrader3.get_aliased_tag("output"), final_name=True))
-
-
-@pytest.mark.parametrize(
-    "nside, error",
-    [
-        (-1, ValueError),
-        (123, ValueError),
-    ],
-)
-def test_ObsCondition_bad_nside(nside, error):
-    """Test bad nside should raise Value and Type errors."""
-    with pytest.raises(error):
-        ObsCondition.make_stage(nside=nside)
-
-
-@pytest.mark.parametrize(
-    "mask, error",
-    [
-        ("xx", ValueError),
-        ("", ValueError),
-    ],
-)
-def test_ObsCondition_bad_mask(mask, error):
-    """Test bad mask should raise Value and Type errors."""
-    with pytest.raises(error):
-        ObsCondition.make_stage(mask=mask)
-
-
-@pytest.mark.parametrize(
-    "weight, error",
-    [
-        ("xx", ValueError),
-    ],
-)
-def test_ObsCondition_bad_weight(weight, error):
-    """Test bad weight should raise Value and Type errors."""
-    with pytest.raises(error):
-        ObsCondition.make_stage(weight=weight)
-
-
-@pytest.mark.parametrize(
-    "map_dict, error",
-    [
-        # band-dependent
-        ({"m5": "xx"}, TypeError),
-        ({"m5": {"u": False}}, TypeError),
-        ({"m5": {"u": "xx"}}, ValueError),
-        ({"m5": {}}, ValueError),
-        ({"nVisYr": "xx"}, TypeError),
-        ({"gamma": {"u": "xx"}}, ValueError),
-        ({"msky": {"u": "xx"}}, ValueError),
-        ({"theta": {"u": False}}, TypeError),
-        ({"km": {"u": False}}, TypeError),
-        # band-independent
-        ({"airmass": "xx"}, ValueError),
-        ({"airmass": False}, TypeError),
-        ({"tvis": False}, TypeError),
-        # wrong key name
-        ({"m5sigma": {"u": 27}}, ValueError),
-        # nYrObs not float
-        ({"nYrObs": "xx"}, TypeError),
-    ],
-)
-def test_ObsCondition_bad_map_dict(map_dict, error):
-    """Test bad map_dict that should raise Value and Type errors."""
-    with pytest.raises(error):
-        ObsCondition.make_stage(map_dict=map_dict)
-
-
-def test_ObsCondition_extended(data):
-    # Testing extended parameter values
-    weight = ""
-    map_dict = {
-        "airmass": os.path.join(
-            os.path.dirname(__file__),
-            "../../src/rail/examples_data/creation_data/data/survey_conditions/minion_1016_dc2_Median_airmass_i_and_nightlt1825_HEAL.fits",
-        ),
-        "EBV": 0.0,
-        "nVisYr": {"u": 50.0},
-        "tvis": 30.0,
-    }
-    tot_nVis_flag = True
-    random_seed = None
-
-    degrader_ext = ObsCondition.make_stage(
-        weight=weight,
-        tot_nVis_flag=tot_nVis_flag,
-        random_seed=random_seed,
-        map_dict=map_dict,
-    )
-    degrader_ext(data)
-    degrader_ext.__repr__()
-
-    os.remove(degrader_ext.get_output(degrader_ext.get_aliased_tag("output"), final_name=True))
-
-
-def test_ObsCondition_empty_map_dict(data):
-    """Test control with random seeds."""
-    degrader1 = ObsCondition.make_stage(random_seed=0, map_dict={})
-    degrader2 = LsstErrorModel()
-
-    # make sure setting the same seeds yields the same output
-    degraded_data1 = degrader1(data).data
-    degraded_data2 = degrader2(data.data, random_state=0)
-    assert degraded_data1.equals(degraded_data2)
-
-    os.remove(degrader1.get_output(degrader1.get_aliased_tag("output"), final_name=True))
-    
-    
-def test_ObsCondition_renameDict(data):
-    """Test with no renameDict included"""
-    degrader1 = ObsCondition.make_stage(random_seed=0, map_dict={"EBV": 0.0,"renameDict": {"u": "u"},})
-
-    # make sure setting the same seeds yields the same output
-    degraded_data1 = degrader1(data).data
-
-    os.remove(degrader1.get_output(degrader1.get_aliased_tag("output"), final_name=True))
-
